PD-8695exrjm : Menu component added to all pages
