+++ web/src/pages/Contact/Contact.tsx
@@ -1,7 +1,14 @@
+import Menu from '../../components/Menu/Menu';
 import './Contact.module.css';
 
 const Contact = () => {
-    return <div> Contact</div>;
+    return (
+        <div>
+
+            <Menu />
+            <div> Contact</div>
+        </div>
+    );
 };
 
 export default Contact;
