+++ web/src/pages/Media/Media.tsx
@@ -1,7 +1,13 @@
+import Menu from '../../components/Menu/Menu';
 import './Media.module.css';
 
 const Media = () => {
-    return <div>Media</div>;
+    return(
+        <>
+            <Menu />
+            <div>Media</div>
+        </>
+    );
 };
 
 export default Media;
