+++ web/src/pages/Reviews/Reviews.tsx
@@ -1,7 +1,14 @@
+import Menu from '../../components/Menu/Menu';
 import './Reviews.module.css';
 
 const Reviews = () => {
-    return <div>Reviews</div>;
+    return (
+        <div>
+            <Menu />
+            <div>Reviews</div>
+
+        </div>
+    );
 };
 
 export default Reviews;
